+++ CrimeMonitor/CrimeMonitor.cs
@@ -1057,10 +1057,15 @@ namespace EddiCrimeMonitor
                 // Filter stations within the faction system which meet the station type prioritization,
                 // max distance from the main star, game version, and landing pad size requirements
                 LandingPadSize padSize = EDDI.Instance?.CurrentShip?.Size ?? LandingPadSize.Large;
-                List<Station> factionStations = !prioritizeOrbitalStations && (EDDI.Instance?.inHorizons ?? false) ? factionStarSystem.stations : factionStarSystem.orbitalstations
-                    .Where(s => s.stationservices.Count > 0).ToList();
-                factionStations = factionStations.Where(s => s.distancefromstar <= maxStationDistanceFromStarLs).ToList();
-                factionStations = factionStations.Where(s => s.LandingPadCheck(padSize)).ToList();
+                List<Station> factionStations = !prioritizeOrbitalStations && (EDDI.Instance?.inHorizons ?? false)
+                    ? factionStarSystem.stations
+                    : factionStarSystem.orbitalstations;
+                factionStations = factionStations
+                    .Where(s => s.Model != StationModel.FleetCarrier)
+                    .Where(s => s.stationservices.Count > 0)
+                    .Where(s => s.distancefromstar <= maxStationDistanceFromStarLs)
+                    .Where(s => s.LandingPadCheck(padSize))
+                    .ToList();
 
                 // Build list to find the faction station nearest to the main star
                 SortedList<decimal, string> nearestList = new SortedList<decimal, string>();